Track shader header dependencies and recompile all shaders if one of its headers has changed.

+++ fancy_core/Common/MathUtil.h
@@ -131,7 +131,7 @@ namespace Fancy { namespace MathUtil {
       i /= aBase;
     }
 
-    return r;
+    return (float) r;
   }
   //---------------------------------------------------------------------------//
   struct Hasher
